requirements.txt
pystray
pyperclip
pyautogui
Pillow
pynput
